put LegistarSite in new file; attempt to import
