+++ civic_scraper/platforms/__init__.py
@@ -1,1 +1,2 @@
 from .civic_plus.site import Site as CivicPlusSite
+from .legistar.scraper import LegistarSite
